+++ CHANGELOG.md
@@ -223,6 +223,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 ### Added
 
 - Implement the following traits
+  - AsMut\<BerRef\> for Ber
   - AsRef\<BerRef\> for Ber
   - From\<bool\> for Ber
   - From\<T: PrimInt\> for Ber (`PrimInt` is `i8` or `u8` or ... `u128` or `isize` or `usize`.)